refactor: Consolidate post rendering and utilities
Extracts the rendering logic from art (`/l/*`) and writing (`/w/*`) slug
pages into a single reusable `PostRenderer` component. This
significantly reduces code duplication.
- Centralizes the `formatDate` function into `lib/utils.ts` and removes
the redundant per-directory utility files.
- Updates the slug pages to use the new component and centralized
utility, simplifying their implementation.
