--- app/l/utils.ts
@@ -1,45 +1,0 @@
-type Metadata = {
-    title: string
-    publishedAt: string
-    summary: string
-    image?: string
-    category?: string
-    languages?: string
-    tags?: string
-}
-
-export function formatDate(date: string, includeRelative = false) {
-    let currentDate = new Date()
-    if (!date.includes('T')) {
-        date = `${date}T00:00:00`
-    }
-    let targetDate = new Date(date)
-
-    let yearsAgo = currentDate.getFullYear() - targetDate.getFullYear()
-    let monthsAgo = currentDate.getMonth() - targetDate.getMonth()
-    let daysAgo = currentDate.getDate() - targetDate.getDate()
-
-    let formattedDate = ''
-
-    if (yearsAgo > 0) {
-        formattedDate = `${yearsAgo}y ago`
-    } else if (monthsAgo > 0) {
-        formattedDate = `${monthsAgo}mo ago`
-    } else if (daysAgo > 0) {
-        formattedDate = `${daysAgo}d ago`
-    } else {
-        formattedDate = 'Today'
-    }
-
-    let fullDate = targetDate.toLocaleString('en-us', {
-        month: 'short',
-        day: 'numeric',
-        year: 'numeric',
-    })
-
-    if (!includeRelative) {
-        return fullDate
-    }
-
-    return `${fullDate} (${formattedDate})`
-}
